requirements.txt
faker
mysql-connector-python
sqlalchemy
python-dotenv